Unit tests and some fixes for Rectangle
