+++ src/main/java/org/codetome/hexameter/core/api/Hexagon.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.codetome.hexameter.core.api.contract.SatelliteData;
 import org.codetome.hexameter.core.backport.Optional;
-import org.codetome.hexameter.core.internal.impl.Rectangle;
 
 /**
  * Represents a Hexagon.
